Refactor OpenStack cloud provider
1.split loadbalancer and volume by version
2.extract Zones's function into openstack_zones.go
3.simple openstack.go by the Interface of cloud.go
4.extract common functions into common util file
