+++ pkg/cloudprovider/providers/openstack/openstack_instance_metadata.go
