+++ pkg/cloudprovider/providers/openstack/openstack_instance_metadata_test.go
