+++ pkg/cloudprovider/providers/openstack/openstack_loadbalancer_v2.go
@@ -19,17 +19,9 @@ package openstack
 import (
 	"fmt"
 	"net"
-	"strings"
-	"time"
 
 	"github.com/golang/glog"
-	"github.com/gophercloud/gophercloud"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/layer3/floatingips"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas/members"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas/monitors"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas/pools"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas/vips"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas_v2/listeners"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas_v2/loadbalancers"
 	v2monitors "github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/lbaas_v2/monitors"
@@ -37,48 +29,11 @@ import (
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/security/groups"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/security/rules"
 	neutronports "github.com/gophercloud/gophercloud/openstack/networking/v2/ports"
-	"github.com/gophercloud/gophercloud/pagination"
 
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/kubernetes/pkg/api/v1/service"
 	"k8s.io/kubernetes/pkg/cloudprovider"
 )
-
-// Note: when creating a new Loadbalancer (VM), it can take some time before it is ready for use,
-// this timeout is used for waiting until the Loadbalancer provisioning status goes to ACTIVE state.
-const (
-	// loadbalancerActive* is configuration of exponential backoff for
-	// going into ACTIVE loadbalancer provisioning status. Starting with 1
-	// seconds, multiplying by 1.2 with each step and taking 19 steps at maximum
-	// it will time out after 128s, which roughly corresponds to 120s
-	loadbalancerActiveInitDealy = 1 * time.Second
-	loadbalancerActiveFactor    = 1.2
-	loadbalancerActiveSteps     = 19
-
-	// loadbalancerDelete* is configuration of exponential backoff for
-	// waiting for delete operation to complete. Starting with 1
-	// seconds, multiplying by 1.2 with each step and taking 13 steps at maximum
-	// it will time out after 32s, which roughly corresponds to 30s
-	loadbalancerDeleteInitDealy = 1 * time.Second
-	loadbalancerDeleteFactor    = 1.2
-	loadbalancerDeleteSteps     = 13
-
-	activeStatus = "ACTIVE"
-	errorStatus  = "ERROR"
-
-	ServiceAnnotationLoadBalancerFloatingNetworkId = "loadbalancer.openstack.org/floating-network-id"
-
-	// ServiceAnnotationLoadBalancerInternal is the annotation used on the service
-	// to indicate that we want an internal loadbalancer service.
-	// If the value of ServiceAnnotationLoadBalancerInternal is false, it indicates that we want an external loadbalancer service. Default to true.
-	ServiceAnnotationLoadBalancerInternal = "service.beta.kubernetes.io/openstack-internal-load-balancer"
-)
-
-// LoadBalancer implementation for LBaaS v1
-type LbaasV1 struct {
-	LoadBalancer
-}
 
 // LoadBalancer implementation for LBaaS v2
 type LbaasV2 struct {
@@ -86,425 +41,6 @@ type LbaasV2 struct {
 }
 
 type empty struct{}
-
-func networkExtensions(client *gophercloud.ServiceClient) (map[string]bool, error) {
-	seen := make(map[string]bool)
-
-	pager := extensions.List(client)
-	err := pager.EachPage(func(page pagination.Page) (bool, error) {
-		exts, err := extensions.ExtractExtensions(page)
-		if err != nil {
-			return false, err
-		}
-		for _, ext := range exts {
-			seen[ext.Alias] = true
-		}
-		return true, nil
-	})
-
-	return seen, err
-}
-
-func getFloatingIPByPortID(client *gophercloud.ServiceClient, portID string) (*floatingips.FloatingIP, error) {
-	opts := floatingips.ListOpts{
-		PortID: portID,
-	}
-	pager := floatingips.List(client, opts)
-
-	floatingIPList := make([]floatingips.FloatingIP, 0, 1)
-
-	err := pager.EachPage(func(page pagination.Page) (bool, error) {
-		f, err := floatingips.ExtractFloatingIPs(page)
-		if err != nil {
-			return false, err
-		}
-		floatingIPList = append(floatingIPList, f...)
-		if len(floatingIPList) > 1 {
-			return false, ErrMultipleResults
-		}
-		return true, nil
-	})
-	if err != nil {
-		if isNotFound(err) {
-			return nil, ErrNotFound
-		}
-		return nil, err
-	}
-
-	if len(floatingIPList) == 0 {
-		return nil, ErrNotFound
-	} else if len(floatingIPList) > 1 {
-		return nil, ErrMultipleResults
-	}
-
-	return &floatingIPList[0], nil
-}
-
-func getPoolByName(client *gophercloud.ServiceClient, name string) (*pools.Pool, error) {
-	opts := pools.ListOpts{
-		Name: name,
-	}
-	pager := pools.List(client, opts)
-
-	poolList := make([]pools.Pool, 0, 1)
-
-	err := pager.EachPage(func(page pagination.Page) (bool, error) {
-		p, err := pools.ExtractPools(page)
-		if err != nil {
-			return false, err
-		}
-		poolList = append(poolList, p...)
-		if len(poolList) > 1 {
-			return false, ErrMultipleResults
-		}
-		return true, nil
-	})
-	if err != nil {
-		if isNotFound(err) {
-			return nil, ErrNotFound
-		}
-		return nil, err
-	}
-
-	if len(poolList) == 0 {
-		return nil, ErrNotFound
-	} else if len(poolList) > 1 {
-		return nil, ErrMultipleResults
-	}
-
-	return &poolList[0], nil
-}
-
-func getVipByName(client *gophercloud.ServiceClient, name string) (*vips.VirtualIP, error) {
-	opts := vips.ListOpts{
-		Name: name,
-	}
-	pager := vips.List(client, opts)
-
-	vipList := make([]vips.VirtualIP, 0, 1)
-
-	err := pager.EachPage(func(page pagination.Page) (bool, error) {
-		v, err := vips.ExtractVIPs(page)
-		if err != nil {
-			return false, err
-		}
-		vipList = append(vipList, v...)
-		if len(vipList) > 1 {
-			return false, ErrMultipleResults
-		}
-		return true, nil
-	})
-	if err != nil {
-		if isNotFound(err) {
-			return nil, ErrNotFound
-		}
-		return nil, err
-	}
-
-	if len(vipList) == 0 {
-		return nil, ErrNotFound
-	} else if len(vipList) > 1 {
-		return nil, ErrMultipleResults
-	}
-
-	return &vipList[0], nil
-}
-
-func getLoadbalancerByName(client *gophercloud.ServiceClient, name string) (*loadbalancers.LoadBalancer, error) {
-	opts := loadbalancers.ListOpts{
-		Name: name,
-	}
-	pager := loadbalancers.List(client, opts)
-
-	loadbalancerList := make([]loadbalancers.LoadBalancer, 0, 1)
-
-	err := pager.EachPage(func(page pagination.Page) (bool, error) {
-		v, err := loadbalancers.ExtractLoadBalancers(page)
-		if err != nil {
-			return false, err
-		}
-		loadbalancerList = append(loadbalancerList, v...)
-		if len(loadbalancerList) > 1 {
-			return false, ErrMultipleResults
-		}
-		return true, nil
-	})
-	if err != nil {
-		if isNotFound(err) {
-			return nil, ErrNotFound
-		}
-		return nil, err
-	}
-
-	if len(loadbalancerList) == 0 {
-		return nil, ErrNotFound
-	} else if len(loadbalancerList) > 1 {
-		return nil, ErrMultipleResults
-	}
-
-	return &loadbalancerList[0], nil
-}
-
-func getListenersByLoadBalancerID(client *gophercloud.ServiceClient, id string) ([]listeners.Listener, error) {
-	var existingListeners []listeners.Listener
-	err := listeners.List(client, listeners.ListOpts{LoadbalancerID: id}).EachPage(func(page pagination.Page) (bool, error) {
-		listenerList, err := listeners.ExtractListeners(page)
-		if err != nil {
-			return false, err
-		}
-		for _, l := range listenerList {
-			for _, lb := range l.Loadbalancers {
-				if lb.ID == id {
-					existingListeners = append(existingListeners, l)
-					break
-				}
-			}
-		}
-
-		return true, nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return existingListeners, nil
-}
-
-// get listener for a port or nil if does not exist
-func getListenerForPort(existingListeners []listeners.Listener, port v1.ServicePort) *listeners.Listener {
-	for _, l := range existingListeners {
-		if listeners.Protocol(l.Protocol) == toListenersProtocol(port.Protocol) && l.ProtocolPort == int(port.Port) {
-			return &l
-		}
-	}
-
-	return nil
-}
-
-// Get pool for a listener. A listener always has exactly one pool.
-func getPoolByListenerID(client *gophercloud.ServiceClient, loadbalancerID string, listenerID string) (*v2pools.Pool, error) {
-	listenerPools := make([]v2pools.Pool, 0, 1)
-	err := v2pools.List(client, v2pools.ListOpts{LoadbalancerID: loadbalancerID}).EachPage(func(page pagination.Page) (bool, error) {
-		poolsList, err := v2pools.ExtractPools(page)
-		if err != nil {
-			return false, err
-		}
-		for _, p := range poolsList {
-			for _, l := range p.Listeners {
-				if l.ID == listenerID {
-					listenerPools = append(listenerPools, p)
-				}
-			}
-		}
-		if len(listenerPools) > 1 {
-			return false, ErrMultipleResults
-		}
-		return true, nil
-	})
-	if err != nil {
-		if isNotFound(err) {
-			return nil, ErrNotFound
-		}
-		return nil, err
-	}
-
-	if len(listenerPools) == 0 {
-		return nil, ErrNotFound
-	} else if len(listenerPools) > 1 {
-		return nil, ErrMultipleResults
-	}
-
-	return &listenerPools[0], nil
-}
-
-func getMembersByPoolID(client *gophercloud.ServiceClient, id string) ([]v2pools.Member, error) {
-	var members []v2pools.Member
-	err := v2pools.ListMembers(client, id, v2pools.ListMembersOpts{}).EachPage(func(page pagination.Page) (bool, error) {
-		membersList, err := v2pools.ExtractMembers(page)
-		if err != nil {
-			return false, err
-		}
-		members = append(members, membersList...)
-
-		return true, nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return members, nil
-}
-
-// Check if a member exists for node
-func memberExists(members []v2pools.Member, addr string, port int) bool {
-	for _, member := range members {
-		if member.Address == addr && member.ProtocolPort == port {
-			return true
-		}
-	}
-
-	return false
-}
-
-func popListener(existingListeners []listeners.Listener, id string) []listeners.Listener {
-	for i, existingListener := range existingListeners {
-		if existingListener.ID == id {
-			existingListeners[i] = existingListeners[len(existingListeners)-1]
-			existingListeners = existingListeners[:len(existingListeners)-1]
-			break
-		}
-	}
-
-	return existingListeners
-}
-
-func popMember(members []v2pools.Member, addr string, port int) []v2pools.Member {
-	for i, member := range members {
-		if member.Address == addr && member.ProtocolPort == port {
-			members[i] = members[len(members)-1]
-			members = members[:len(members)-1]
-		}
-	}
-
-	return members
-}
-
-func getSecurityGroupName(clusterName string, service *v1.Service) string {
-	return fmt.Sprintf("lb-sg-%s-%v", clusterName, service.Name)
-}
-
-func getSecurityGroupRules(client *gophercloud.ServiceClient, opts rules.ListOpts) ([]rules.SecGroupRule, error) {
-
-	pager := rules.List(client, opts)
-
-	var securityRules []rules.SecGroupRule
-
-	err := pager.EachPage(func(page pagination.Page) (bool, error) {
-		ruleList, err := rules.ExtractRules(page)
-		if err != nil {
-			return false, err
-		}
-		securityRules = append(securityRules, ruleList...)
-		return true, nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return securityRules, nil
-}
-
-func waitLoadbalancerActiveProvisioningStatus(client *gophercloud.ServiceClient, loadbalancerID string) (string, error) {
-	backoff := wait.Backoff{
-		Duration: loadbalancerActiveInitDealy,
-		Factor:   loadbalancerActiveFactor,
-		Steps:    loadbalancerActiveSteps,
-	}
-
-	var provisioningStatus string
-	err := wait.ExponentialBackoff(backoff, func() (bool, error) {
-		loadbalancer, err := loadbalancers.Get(client, loadbalancerID).Extract()
-		if err != nil {
-			return false, err
-		}
-		provisioningStatus = loadbalancer.ProvisioningStatus
-		if loadbalancer.ProvisioningStatus == activeStatus {
-			return true, nil
-		} else if loadbalancer.ProvisioningStatus == errorStatus {
-			return true, fmt.Errorf("Loadbalancer has gone into ERROR state")
-		} else {
-			return false, nil
-		}
-
-	})
-
-	if err == wait.ErrWaitTimeout {
-		err = fmt.Errorf("Loadbalancer failed to go into ACTIVE provisioning status within alloted time")
-	}
-	return provisioningStatus, err
-}
-
-func waitLoadbalancerDeleted(client *gophercloud.ServiceClient, loadbalancerID string) error {
-	backoff := wait.Backoff{
-		Duration: loadbalancerDeleteInitDealy,
-		Factor:   loadbalancerDeleteFactor,
-		Steps:    loadbalancerDeleteSteps,
-	}
-	err := wait.ExponentialBackoff(backoff, func() (bool, error) {
-		_, err := loadbalancers.Get(client, loadbalancerID).Extract()
-		if err != nil {
-			if err == ErrNotFound {
-				return true, nil
-			} else {
-				return false, err
-			}
-		} else {
-			return false, nil
-		}
-	})
-
-	if err == wait.ErrWaitTimeout {
-		err = fmt.Errorf("Loadbalancer failed to delete within the alloted time")
-	}
-
-	return err
-}
-
-func toRuleProtocol(protocol v1.Protocol) rules.RuleProtocol {
-	switch protocol {
-	case v1.ProtocolTCP:
-		return rules.ProtocolTCP
-	case v1.ProtocolUDP:
-		return rules.ProtocolUDP
-	default:
-		return rules.RuleProtocol(strings.ToLower(string(protocol)))
-	}
-}
-
-func toListenersProtocol(protocol v1.Protocol) listeners.Protocol {
-	switch protocol {
-	case v1.ProtocolTCP:
-		return listeners.ProtocolTCP
-	default:
-		return listeners.Protocol(string(protocol))
-	}
-}
-
-func createNodeSecurityGroup(client *gophercloud.ServiceClient, nodeSecurityGroupID string, port int, protocol v1.Protocol, lbSecGroup string) error {
-	v4NodeSecGroupRuleCreateOpts := rules.CreateOpts{
-		Direction:     rules.DirIngress,
-		PortRangeMax:  port,
-		PortRangeMin:  port,
-		Protocol:      toRuleProtocol(protocol),
-		RemoteGroupID: lbSecGroup,
-		SecGroupID:    nodeSecurityGroupID,
-		EtherType:     rules.EtherType4,
-	}
-
-	v6NodeSecGroupRuleCreateOpts := rules.CreateOpts{
-		Direction:     rules.DirIngress,
-		PortRangeMax:  port,
-		PortRangeMin:  port,
-		Protocol:      toRuleProtocol(protocol),
-		RemoteGroupID: lbSecGroup,
-		SecGroupID:    nodeSecurityGroupID,
-		EtherType:     rules.EtherType6,
-	}
-
-	_, err := rules.Create(client, v4NodeSecGroupRuleCreateOpts).Extract()
-
-	if err != nil {
-		return err
-	}
-
-	_, err = rules.Create(client, v6NodeSecGroupRuleCreateOpts).Extract()
-
-	if err != nil {
-		return err
-	}
-	return nil
-}
 
 func (lbaas *LbaasV2) createLoadBalancer(service *v1.Service, name string, internalAnnotation bool) (*loadbalancers.LoadBalancer, error) {
 	createOpts := loadbalancers.CreateOpts{
@@ -539,68 +75,6 @@ func (lbaas *LbaasV2) GetLoadBalancer(clusterName string, service *v1.Service) (
 	status.Ingress = []v1.LoadBalancerIngress{{IP: loadbalancer.VipAddress}}
 
 	return status, true, err
-}
-
-// The LB needs to be configured with instance addresses on the same
-// subnet as the LB (aka opts.SubnetId).  Currently we're just
-// guessing that the node's InternalIP is the right address - and that
-// should be sufficient for all "normal" cases.
-func nodeAddressForLB(node *v1.Node) (string, error) {
-	addrs := node.Status.Addresses
-	if len(addrs) == 0 {
-		return "", ErrNoAddressFound
-	}
-
-	for _, addr := range addrs {
-		if addr.Type == v1.NodeInternalIP {
-			return addr.Address, nil
-		}
-	}
-
-	return addrs[0].Address, nil
-}
-
-//getStringFromServiceAnnotation searches a given v1.Service for a specific annotationKey and either returns the annotation's value or a specified defaultSetting
-func getStringFromServiceAnnotation(service *v1.Service, annotationKey string, defaultSetting string) string {
-	glog.V(4).Infof("getStringFromServiceAnnotation(%v, %v, %v)", service, annotationKey, defaultSetting)
-	if annotationValue, ok := service.Annotations[annotationKey]; ok {
-		//if there is an annotation for this setting, set the "setting" var to it
-		// annotationValue can be empty, it is working as designed
-		// it makes possible for instance provisioning loadbalancer without floatingip
-		glog.V(4).Infof("Found a Service Annotation: %v = %v", annotationKey, annotationValue)
-		return annotationValue
-	}
-	//if there is no annotation, set "settings" var to the value from cloud config
-	glog.V(4).Infof("Could not find a Service Annotation; falling back on cloud-config setting: %v = %v", annotationKey, defaultSetting)
-	return defaultSetting
-}
-
-// getSubnetIDForLB returns subnet-id for a specific node
-func getSubnetIDForLB(compute *gophercloud.ServiceClient, node v1.Node) (string, error) {
-	ipAddress, err := nodeAddressForLB(&node)
-	if err != nil {
-		return "", err
-	}
-
-	instanceID := node.Spec.ProviderID
-	if ind := strings.LastIndex(instanceID, "/"); ind >= 0 {
-		instanceID = instanceID[(ind + 1):]
-	}
-
-	interfaces, err := getAttachedInterfacesByID(compute, instanceID)
-	if err != nil {
-		return "", err
-	}
-
-	for _, intf := range interfaces {
-		for _, fixedIP := range intf.FixedIPs {
-			if fixedIP.IPAddress == ipAddress {
-				return intf.NetID, nil
-			}
-		}
-	}
-
-	return "", ErrNotFound
 }
 
 // TODO: This code currently ignores 'region' and always creates a
@@ -1270,346 +744,6 @@ func (lbaas *LbaasV2) EnsureLoadBalancerDeleted(clusterName string, service *v1.
 			if res.Err != nil && !isNotFound(res.Err) {
 				glog.V(1).Infof("Error occurred deleting security group rule: %s: %v", rule.ID, res.Err)
 			}
-		}
-	}
-
-	return nil
-}
-
-func (lb *LbaasV1) GetLoadBalancer(clusterName string, service *v1.Service) (*v1.LoadBalancerStatus, bool, error) {
-	loadBalancerName := cloudprovider.GetLoadBalancerName(service)
-	vip, err := getVipByName(lb.network, loadBalancerName)
-	if err == ErrNotFound {
-		return nil, false, nil
-	}
-	if vip == nil {
-		return nil, false, err
-	}
-
-	status := &v1.LoadBalancerStatus{}
-	status.Ingress = []v1.LoadBalancerIngress{{IP: vip.Address}}
-
-	return status, true, err
-}
-
-// TODO: This code currently ignores 'region' and always creates a
-// loadbalancer in only the current OpenStack region.  We should take
-// a list of regions (from config) and query/create loadbalancers in
-// each region.
-
-func (lb *LbaasV1) EnsureLoadBalancer(clusterName string, apiService *v1.Service, nodes []*v1.Node) (*v1.LoadBalancerStatus, error) {
-	glog.V(4).Infof("EnsureLoadBalancer(%v, %v, %v, %v, %v, %v, %v)", clusterName, apiService.Namespace, apiService.Name, apiService.Spec.LoadBalancerIP, apiService.Spec.Ports, nodes, apiService.Annotations)
-
-	if len(lb.opts.SubnetId) == 0 {
-		// Get SubnetId automatically.
-		// The LB needs to be configured with instance addresses on the same subnet, so get SubnetId by one node.
-		subnetID, err := getSubnetIDForLB(lb.compute, *nodes[0])
-		if err != nil {
-			glog.Warningf("Failed to find subnet-id for loadbalancer service %s/%s: %v", apiService.Namespace, apiService.Name, err)
-			return nil, fmt.Errorf("No subnet-id for service %s/%s : subnet-id not set in cloud provider config, "+
-				"and failed to find subnet-id from OpenStack: %v", apiService.Namespace, apiService.Name, err)
-		}
-		lb.opts.SubnetId = subnetID
-	}
-
-	floatingPool := getStringFromServiceAnnotation(apiService, ServiceAnnotationLoadBalancerFloatingNetworkId, lb.opts.FloatingNetworkId)
-	glog.V(4).Infof("EnsureLoadBalancer using floatingPool: %v", floatingPool)
-
-	var internalAnnotation bool
-	internal := getStringFromServiceAnnotation(apiService, ServiceAnnotationLoadBalancerInternal, "true")
-	switch internal {
-	case "true":
-		glog.V(4).Infof("Ensure an internal loadbalancer service.")
-		internalAnnotation = true
-	case "false":
-		if len(floatingPool) != 0 {
-			glog.V(4).Infof("Ensure an external loadbalancer service.")
-			internalAnnotation = false
-		} else {
-			return nil, fmt.Errorf("floating-network-id or loadbalancer.openstack.org/floating-network-id should be specified when service.beta.kubernetes.io/openstack-internal-load-balancer is false")
-		}
-	default:
-		return nil, fmt.Errorf("unknow service.beta.kubernetes.io/openstack-internal-load-balancer annotation: %v, specify \"true\" or \"false\".",
-			internal)
-	}
-
-	ports := apiService.Spec.Ports
-	if len(ports) > 1 {
-		return nil, fmt.Errorf("multiple ports are not supported in openstack v1 load balancers")
-	} else if len(ports) == 0 {
-		return nil, fmt.Errorf("no ports provided to openstack load balancer")
-	}
-
-	// The service controller verified all the protocols match on the ports, just check and use the first one
-	// TODO: Convert all error messages to use an event recorder
-	if ports[0].Protocol != v1.ProtocolTCP {
-		return nil, fmt.Errorf("Only TCP LoadBalancer is supported for openstack load balancers")
-	}
-
-	affinity := apiService.Spec.SessionAffinity
-	var persistence *vips.SessionPersistence
-	switch affinity {
-	case v1.ServiceAffinityNone:
-		persistence = nil
-	case v1.ServiceAffinityClientIP:
-		persistence = &vips.SessionPersistence{Type: "SOURCE_IP"}
-	default:
-		return nil, fmt.Errorf("unsupported load balancer affinity: %v", affinity)
-	}
-
-	sourceRanges, err := service.GetLoadBalancerSourceRanges(apiService)
-	if err != nil {
-		return nil, err
-	}
-
-	if !service.IsAllowAll(sourceRanges) {
-		return nil, fmt.Errorf("Source range restrictions are not supported for openstack load balancers")
-	}
-
-	glog.V(2).Infof("Checking if openstack load balancer already exists: %s", cloudprovider.GetLoadBalancerName(apiService))
-	_, exists, err := lb.GetLoadBalancer(clusterName, apiService)
-	if err != nil {
-		return nil, fmt.Errorf("error checking if openstack load balancer already exists: %v", err)
-	}
-
-	// TODO: Implement a more efficient update strategy for common changes than delete & create
-	// In particular, if we implement hosts update, we can get rid of UpdateHosts
-	if exists {
-		err := lb.EnsureLoadBalancerDeleted(clusterName, apiService)
-		if err != nil {
-			return nil, fmt.Errorf("error deleting existing openstack load balancer: %v", err)
-		}
-	}
-
-	lbmethod := pools.LBMethod(lb.opts.LBMethod)
-	if lbmethod == "" {
-		lbmethod = pools.LBMethodRoundRobin
-	}
-	name := cloudprovider.GetLoadBalancerName(apiService)
-	pool, err := pools.Create(lb.network, pools.CreateOpts{
-		Name:     name,
-		Protocol: pools.ProtocolTCP,
-		SubnetID: lb.opts.SubnetId,
-		LBMethod: lbmethod,
-	}).Extract()
-	if err != nil {
-		return nil, fmt.Errorf("Error creating pool for openstack load balancer %s: %v", name, err)
-	}
-
-	for _, node := range nodes {
-		addr, err := nodeAddressForLB(node)
-		if err != nil {
-			return nil, err
-		}
-
-		_, err = members.Create(lb.network, members.CreateOpts{
-			PoolID:       pool.ID,
-			ProtocolPort: int(ports[0].NodePort), //Note: only handles single port
-			Address:      addr,
-		}).Extract()
-		if err != nil {
-			return nil, fmt.Errorf("Error creating member for the pool(%s) of openstack load balancer %s: %v",
-				pool.ID, name, err)
-		}
-	}
-
-	var mon *monitors.Monitor
-	if lb.opts.CreateMonitor {
-		mon, err = monitors.Create(lb.network, monitors.CreateOpts{
-			Type:       monitors.TypeTCP,
-			Delay:      int(lb.opts.MonitorDelay.Duration.Seconds()),
-			Timeout:    int(lb.opts.MonitorTimeout.Duration.Seconds()),
-			MaxRetries: int(lb.opts.MonitorMaxRetries),
-		}).Extract()
-		if err != nil {
-			return nil, fmt.Errorf("Error creating monitor for openstack load balancer %s: %v", name, err)
-		}
-
-		_, err = pools.AssociateMonitor(lb.network, pool.ID, mon.ID).Extract()
-		if err != nil {
-			return nil, fmt.Errorf("Error associating monitor(%s) with pool(%s) for"+
-				"openstack load balancer %s: %v", mon.ID, pool.ID, name, err)
-		}
-	}
-
-	createOpts := vips.CreateOpts{
-		Name:         name,
-		Description:  fmt.Sprintf("Kubernetes external service %s", name),
-		Protocol:     "TCP",
-		ProtocolPort: int(ports[0].Port), //TODO: need to handle multi-port
-		PoolID:       pool.ID,
-		SubnetID:     lb.opts.SubnetId,
-		Persistence:  persistence,
-	}
-
-	loadBalancerIP := apiService.Spec.LoadBalancerIP
-	if loadBalancerIP != "" && internalAnnotation {
-		createOpts.Address = loadBalancerIP
-	}
-
-	vip, err := vips.Create(lb.network, createOpts).Extract()
-	if err != nil {
-		return nil, fmt.Errorf("Error creating vip for openstack load balancer %s: %v", name, err)
-	}
-
-	status := &v1.LoadBalancerStatus{}
-
-	status.Ingress = []v1.LoadBalancerIngress{{IP: vip.Address}}
-
-	if floatingPool != "" && !internalAnnotation {
-		floatIPOpts := floatingips.CreateOpts{
-			FloatingNetworkID: floatingPool,
-			PortID:            vip.PortID,
-		}
-
-		loadBalancerIP := apiService.Spec.LoadBalancerIP
-		if loadBalancerIP != "" {
-			floatIPOpts.FloatingIP = loadBalancerIP
-		}
-
-		floatIP, err := floatingips.Create(lb.network, floatIPOpts).Extract()
-		if err != nil {
-			return nil, fmt.Errorf("Error creating floatingip for openstack load balancer %s: %v", name, err)
-		}
-
-		status.Ingress = append(status.Ingress, v1.LoadBalancerIngress{IP: floatIP.FloatingIP})
-	}
-
-	return status, nil
-
-}
-
-func (lb *LbaasV1) UpdateLoadBalancer(clusterName string, service *v1.Service, nodes []*v1.Node) error {
-	loadBalancerName := cloudprovider.GetLoadBalancerName(service)
-	glog.V(4).Infof("UpdateLoadBalancer(%v, %v, %v)", clusterName, loadBalancerName, nodes)
-
-	vip, err := getVipByName(lb.network, loadBalancerName)
-	if err != nil {
-		return err
-	}
-
-	// Set of member (addresses) that _should_ exist
-	addrs := map[string]bool{}
-	for _, node := range nodes {
-		addr, err := nodeAddressForLB(node)
-		if err != nil {
-			return err
-		}
-
-		addrs[addr] = true
-	}
-
-	// Iterate over members that _do_ exist
-	pager := members.List(lb.network, members.ListOpts{PoolID: vip.PoolID})
-	err = pager.EachPage(func(page pagination.Page) (bool, error) {
-		memList, err := members.ExtractMembers(page)
-		if err != nil {
-			return false, err
-		}
-
-		for _, member := range memList {
-			if _, found := addrs[member.Address]; found {
-				// Member already exists
-				delete(addrs, member.Address)
-			} else {
-				// Member needs to be deleted
-				err = members.Delete(lb.network, member.ID).ExtractErr()
-				if err != nil {
-					return false, err
-				}
-			}
-		}
-
-		return true, nil
-	})
-	if err != nil {
-		return err
-	}
-
-	// Anything left in addrs is a new member that needs to be added
-	for addr := range addrs {
-		_, err := members.Create(lb.network, members.CreateOpts{
-			PoolID:       vip.PoolID,
-			Address:      addr,
-			ProtocolPort: vip.ProtocolPort,
-		}).Extract()
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (lb *LbaasV1) EnsureLoadBalancerDeleted(clusterName string, service *v1.Service) error {
-	loadBalancerName := cloudprovider.GetLoadBalancerName(service)
-	glog.V(4).Infof("EnsureLoadBalancerDeleted(%v, %v)", clusterName, loadBalancerName)
-
-	vip, err := getVipByName(lb.network, loadBalancerName)
-	if err != nil && err != ErrNotFound {
-		return err
-	}
-
-	if vip != nil && vip.PortID != "" {
-		floatingIP, err := getFloatingIPByPortID(lb.network, vip.PortID)
-		if err != nil && !isNotFound(err) {
-			return err
-		}
-		if floatingIP != nil {
-			err = floatingips.Delete(lb.network, floatingIP.ID).ExtractErr()
-			if err != nil && !isNotFound(err) {
-				return err
-			}
-		}
-	}
-
-	// We have to delete the VIP before the pool can be deleted,
-	// so no point continuing if this fails.
-	if vip != nil {
-		err := vips.Delete(lb.network, vip.ID).ExtractErr()
-		if err != nil && !isNotFound(err) {
-			return err
-		}
-	}
-
-	var pool *pools.Pool
-	if vip != nil {
-		pool, err = pools.Get(lb.network, vip.PoolID).Extract()
-		if err != nil && !isNotFound(err) {
-			return err
-		}
-	} else {
-		// The VIP is gone, but it is conceivable that a Pool
-		// still exists that we failed to delete on some
-		// previous occasion.  Make a best effort attempt to
-		// cleanup any pools with the same name as the VIP.
-		pool, err = getPoolByName(lb.network, service.Name)
-		if err != nil && err != ErrNotFound {
-			return err
-		}
-	}
-
-	if pool != nil {
-		for _, monId := range pool.MonitorIDs {
-			_, err = pools.DisassociateMonitor(lb.network, pool.ID, monId).Extract()
-			if err != nil {
-				return err
-			}
-
-			err = monitors.Delete(lb.network, monId).ExtractErr()
-			if err != nil && !isNotFound(err) {
-				return err
-			}
-		}
-		for _, memberId := range pool.MemberIDs {
-			err = members.Delete(lb.network, memberId).ExtractErr()
-			if err != nil && !isNotFound(err) {
-				return err
-			}
-		}
-		err = pools.Delete(lb.network, pool.ID).ExtractErr()
-		if err != nil && !isNotFound(err) {
-			return err
 		}
 	}
